pyproject.toml
[tool.poetry]
name = "BahaAutoSign"
version = "0.1.0"
description = ""
authors = ["tasi788 <tdc@tdccc.com.tw>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
PyNaCl = "^1.4.0"
pyotp = "^2.6.0"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
